Start moving low-level code as is in register module as to not break the unit tests. Adapt unit tests and imports
