+++ unit_testing/test_register.py
@@ -7,11 +7,10 @@ import logging
 import pytest
 
 # FIXME only during refactor, everything should move to spinalcordtoolbox.module
-from msct_register import *
+from msct_register import register
 from sct_register_to_template import Param
 
 from spinalcordtoolbox.register.register import *
-from spinalcordtoolbox.image import Image
 
 logger = logging.getLogger(__name__)
